Changes to UMD and Typescript support
