+++ knockout.viewmodel.d.ts
@@ -1,0 +1,17 @@
+interface KnockoutViewModelStatic {
+    toModel(viewmodel: any): any;
+    fromModel(model: any, options?: any): any;
+    updateFromModel(viewmodel: any, model: any);
+    
+    // INTERNAL flag: enable logging of conversions
+    // logs will be written to console
+    logging: boolean;
+}
+
+// Extend ko global
+interface KnockoutStatic {
+    viewmodel: KnockoutViewModelStatic;
+}
+
+declare let ko_viewmodel : KnockoutViewModelStatic;
+export = ko_viewmodel;
